separando logica del reduce del componente AddProduct y generando un custom hook para manejar los cambios en el formulario
